requirements.txt
nlg_eval==2.3
numpy==1.20.0
pandas==1.3.5
scikit_learn==1.1.1
six==1.15.0
torch==1.11.0
tqdm==4.49.0
transformers==4.24.0
faiss-gpu==1.7.3